refactor(core): extract qmd-result-resolver seam from orchestrator (#1526 seam 11)
Extract the QMD search-result resolution subsystem (readQmdResultMemory,
resolveColdQmdResultForRecall, storageForAbsoluteQmdResultPath) plus the
two module-level path utilities (qmdCollectionPathParts,
qmdResultPathCandidates) into a new QmdResultResolver coordinator.

Behavior-preserving move: the orchestrator constructs one instance and
keeps thin delegating private methods so all existing call sites
(recallInternal, applyColdFallbackPipeline, loadSearchResultMemoryMap,
suggestLinksForMemory, and the RecallRerankCoordinator callback) work
unchanged.

The coordinator accesses config, the storage router, and the
namespace-resolution helpers through getter callbacks (not captured at
construction), mirroring the RecallRerankCoordinator /
ConversationIndexCoordinator accessor pattern.

Also removes the unnecessary inline-cast pattern for StorageManager.dir
(the getter is typed string) in the moved code.

orchestrator.ts: 18944 -> 18679 LOC (-265).
entity-hardening 246/246 green; tsc clean; lint clean; ratchet DOWN.

Refs #1526
